Organized s3connector into node folder
